- S'ha modificat les variables del DICCP i DICCF per les correctes - Funcions del main.c afegides i configurades

+++ Core/Inc/f2p.h
@@ -3,7 +3,7 @@
 
 #include "main.h"
 
-void DMA2DICCF(DICCF_t *DICCF, uint16_t *buffer);
+void DMA2DICCF(DICCF_t *DICCF, uint32_t *buffer);
 void DIG2DICCF(DICCF_t *DICCF);
 void DICCF2DICCP(DICCF_t *DICCF, DICCP_t *DICCP);
